fix(migration): Use numeric sorting for ID column in pagination
ID column in CSV tables is often text type, causing string sorting
(10001, 10002... before 1806). Cast to int for proper numeric order.

+++ backend/src/routes/migration.routes.ts
@@ -1321,8 +1321,9 @@ router.post('/generate', async (req: Request, res: Response) => {
 
       while (true) {
         // Fetch a batch of records from source (with pagination)
+        // Note: ORDER BY id::int ensures numeric sorting even if id column is text type
         const batchResult = await pools.sourcePool.query(
-          `SELECT id, * FROM public."${table}" ORDER BY id LIMIT $1 OFFSET $2`,
+          `SELECT id, * FROM public."${table}" ORDER BY id::int LIMIT $1 OFFSET $2`,
           [BATCH_FETCH_SIZE, offset]
         );
